% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-analysisGraphic.R
\name{analysis_graphic}
\alias{analysis_graphic}
\title{Bu fonksiyon veri analizi yapmamizi ve gerekli grafikleri cizdirmemizi saglar}
\usage{
analysis_graphic(data)
}
\arguments{
\item{data}{veriyi alacak degisken}
}
\description{
Bu fonksiyon veri analizi yapmamizi ve gerekli grafikleri cizdirmemizi saglar
}
