% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-presence.R
\name{create_sqlite_db}
\alias{create_sqlite_db}
\title{Bu fonksiyon veri setine sqlite formatina donusturur.}
\usage{
create_sqlite_db()
}
\description{
Bu fonksiyon veri setine sqlite formatina donusturur.
}
