% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-integrity.R
\name{data_preprocessing}
\alias{data_preprocessing}
\title{Bu fonksiyon veri on isleme adimi icin kullanilir}
\usage{
data_preprocessing(data)
}
\arguments{
\item{data}{veriyi alacak degisken}
}
\description{
Bu fonksiyon veri on isleme adimi icin kullanilir
}
