% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-write.R
\name{write_data_to_db}
\alias{write_data_to_db}
\title{Bu fonksiyon veri on isleme adimindaki temizlenmis veri setini db ye yazma islevini gerceklestirir}
\usage{
write_data_to_db(data)
}
\arguments{
\item{data}{veriyi alacak degisken}
}
\description{
Bu fonksiyon veri on isleme adimindaki temizlenmis veri setini db ye yazma islevini gerceklestirir
}
